# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SessionStart {
  /** Target role (AI/ML Engineer or Backend Engineer) */
  role: string;
  /** Resume PDF file */
  resume: Blob;
}

export interface InterviewQuestion {
  id: number;
  question: string;
  topic: string;
}

export interface SessionStartResponse {
  session_id: string;
  candidate_name: string;
  role: string;
  extracted_skills: string[];
  questions: InterviewQuestion[];
  total_questions: number;
  experience_level: string;
}

export interface SessionDetail {
  session_id: string;
  candidate_name: string;
  role: string;
  extracted_skills: string[];
  questions: InterviewQuestion[];
  status: string;
  answers_count: number;
  total_questions: number;
  experience_level: string;
}

export interface AnswerInput {
  session_id: string;
  question_id: number;
  answer: string;
}

export interface AnswerResponse {
  success: boolean;
  next_question_index: number;
  is_complete: boolean;
  message: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  strength?: string | null;
  /** @nullable */
  improvement?: string | null;
}

export interface QuestionAnswer {
  question_id: number;
  question: string;
  topic: string;
  /** @nullable */
  answer: string | null;
  answered: boolean;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  strength?: string | null;
  /** @nullable */
  improvement?: string | null;
}

export interface InterviewSummary {
  session_id: string;
  candidate_name: string;
  role: string;
  extracted_skills: string[];
  qa_pairs: QuestionAnswer[];
  topics_covered: string[];
  completion_status: string;
  total_questions: number;
  answered_count: number;
  experience_level: string;
}

export interface ErrorResponse {
  error: string;
}