# lib/api-zod/src/generated/types/answerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerInput {
  session_id: string;
  question_id: number;
  answer: string;
}