# lib/api-zod/src/generated/types/answerResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerResponse {
  success: boolean;
  next_question_index: number;
  is_complete: boolean;
  message: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  strength?: string | null;
  /** @nullable */
  improvement?: string | null;
}