# lib/api-zod/src/generated/types/interviewQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewQuestion {
  id: number;
  question: string;
  topic: string;
}