# lib/api-zod/src/generated/types/interviewSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionAnswer } from './questionAnswer';

export interface InterviewSummary {
  session_id: string;
  candidate_name: string;
  role: string;
  extracted_skills: string[];
  qa_pairs: QuestionAnswer[];
  topics_covered: string[];
  completion_status: string;
  total_questions: number;
  answered_count: number;
  experience_level: string;
}