# lib/api-zod/src/generated/types/questionAnswer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionAnswer {
  question_id: number;
  question: string;
  topic: string;
  /** @nullable */
  answer: string | null;
  answered: boolean;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  strength?: string | null;
  /** @nullable */
  improvement?: string | null;
}