# lib/api-zod/src/generated/types/sessionDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewQuestion } from './interviewQuestion';

export interface SessionDetail {
  session_id: string;
  candidate_name: string;
  role: string;
  extracted_skills: string[];
  questions: InterviewQuestion[];
  status: string;
  answers_count: number;
  total_questions: number;
  experience_level: string;
}