# lib/api-zod/src/generated/types/sessionStart.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-powered interview system API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionStart {
  /** Target role (AI/ML Engineer or Backend Engineer) */
  role: string;
  /** Resume PDF file */
  resume: Blob;
}